created watch,datagather

+++ datagather.py
@@ -89,13 +89,11 @@ def merge_json_results(json1_filename, json2_filename, folder='recalldata'):
         print(f"Deleted {json2_path}")
     else:
         print(f"{json2_path} does not exist.")
-# function to get lat and lon from adress 
-# other helper functions
-# gonna start with active ones (make api call for that)
-# read in apropriate data from API and put it in a dictionary  
 # get ready to export to email
 # read in all active cases currently 
 # write function to loop over 
+# CONCATENATE NEW DATA WHEN IT COMES into json files 
+# make allergy filter and make new notify python file that tells u if 
 
 
 
@@ -119,37 +117,47 @@ def merge_json_results(json1_filename, json2_filename, folder='recalldata'):
 
 
 
-def get_active_cases():
+def get_active_cases(offset:int,folder="recalldata") ->dict:
     
     final_dict={
 
     }
+    
     #get in states, then get nationwide and if id are not the same then add it in to json
-    for key in CONVERSIONDICT.keys():  ## replace with keys or values 
+    for key in list(CONVERSIONDICT.keys())[:-offset]:  
         r=requests.get(f'https://api.fda.gov/food/enforcement.json?search=distribution_pattern:"{key}"+AND+status:"Ongoing"+AND+(classification:"Class II"+classification:"Class I")&limit=500')
         
         if(r.status_code==200):
             recallData=r.json()
-            with open(f'recalldata/{key}.json', 'w') as json_file:
-                    json.dump(recallData, json_file, indent=4)
+
+            final_dict[f"{key}"]=recallData
+
+            with open(f'{folder}/{key}.json', 'w') as json_file1:
+                    json.dump(recallData, json_file1, indent=4)
 
         else: 
             print(f"ERROR: RESPONSE CODE{r.status_code} at {key}")
 
-    for value in CONVERSIONDICT.values():  ## replace with keys or values 
+    for value in list(CONVERSIONDICT.values())[:-offset]: 
             r=requests.get(f'https://api.fda.gov/food/enforcement.json?search=distribution_pattern:"{value}"+AND+status:"Ongoing"+AND+(classification:"Class II"+classification:"Class I")&limit=500')
             
             if(r.status_code==200):
                 recallData=r.json()
-                with open(f'recalldata/{value}.json', 'w') as json_file:
-                        json.dump(recallData, json_file, indent=4)
+
+                final_dict[f"{value}"]=recallData
+
+                with open(f'{folder}/{value}.json', 'w') as json_file2:
+                        json.dump(recallData, json_file2, indent=4)
 
             else: 
                 print(f"ERROR: RESPONSE CODE{r.status_code} at {value}")
 
+    return final_dict
     
 
-def merge_json_results(json1_filename, json2_filename, folder='recalldata'):
+    
+
+def merge_json_results(json1_filename:str, json2_filename:str, makes_files:bool, folder='recalldata'):
     # Construct full file paths
     json1_path = os.path.join(folder, json1_filename)
     json2_path = os.path.join(folder, json2_filename)
@@ -169,30 +177,32 @@ def merge_json_results(json1_filename, json2_filename, folder='recalldata'):
         raise KeyError("One or both JSONs do not contain a 'results' key.")
 
     # Overwrite json1 file with the new json3 content
-    with open(json1_path, 'w') as file1:
-        json.dump(json3, file1, indent=4)
-    
-    print(f"Successfully merged and saved to {json1_path}")
-
-    if os.path.exists(json2_path):
-        os.remove(json2_path)
-        print(f"Deleted {json2_path}")
-    else:
-        print(f"{json2_path} does not exist.")
-
-
-
-def collect_data():  #call this function to collect all data
+    if(makes_files):
+        with open(json1_path, 'w') as file1:
+            json.dump(json3, file1, indent=4)
         
-    get_active_cases()
-    for i in range(0,50):
+        print(f"Successfully merged and saved to {json1_path}")
+
+        if os.path.exists(json2_path):
+            os.remove(json2_path)
+            print(f"Deleted {json2_path}")
+        else:
+            print(f"{json2_path} does not exist.")
+    return json3
+
+
+
+def collect_data(offset:int,folder="recalldata"):  #call this function to collect all data
+        
+    get_active_cases(offset,folder=folder)
+    for i in range(0,50-offset):
         key=list(CONVERSIONDICT.keys())[i]
         value=list(CONVERSIONDICT.values())[i]
-        merge_json_results(f'{key}.json', f'{value}.json')
+        merge_json_results(f'{key}.json', f'{value}.json',makes_files=True,folder=folder)
 
 
 
-
+# collect_data(1)
